Fix unsigned subtraction in GetNumTradeUnitsRemaining
UBSan findings 9-10 (CvTradeClasses.cpp:5420):
'7 - 8 cannot be represented in type uint' followed by
'value 4294967295 changed to -1'.

GetNumTradeRoutesPossible() returns uint, so the int result of
GetNumTradeUnits() is promoted to unsigned and the subtraction wraps
when more trade units exist than route slots (a normal transient state
while a route slot source is lost or a unit is queued). The wrapped
value converts back to the intended -1 on two's complement and every
caller compares <= 0, so behavior was correct; the wrap is defined and
the uint->int conversion is implementation-defined, making this a
portability hazard rather than UB.

Cast the uint operand to int so the subtraction is signed throughout
No behavior change.

+++ CvGameCoreDLL_Expansion2/CvTradeClasses.cpp
@@ -5417,7 +5417,8 @@ int CvPlayerTrade::GetNumTradeUnits(bool bIncludeBeingBuilt) const
 //	--------------------------------------------------------------------------------
 int CvPlayerTrade::GetNumTradeUnitsRemaining (bool bIncludeBeingBuilt)
 {
-	return (GetNumTradeRoutesPossible() - GetNumTradeUnits(bIncludeBeingBuilt));
+	//result may be negative - make sure the subtraction happens in signed arithmetic
+	return ((int)GetNumTradeRoutesPossible() - GetNumTradeUnits(bIncludeBeingBuilt));
 }
 
 //	--------------------------------------------------------------------------------
